Get the commands out to a Trait
