+++ src/Commands/InstallCommand.php
@@ -3,10 +3,13 @@
 namespace Native\Electron\Commands;
 
 use Illuminate\Console\Command;
+use Native\Electron\Traits\Installer;
 use Symfony\Component\Process\Process;
 
 class InstallCommand extends Command
 {
+    use Installer;
+
     protected $signature = 'native:install {--force : Overwrite existing files by default} {--installer=npm : The package installer to use: npm, yarn or pnpm}';
 
     protected $description = 'Install all of the NativePHP resources';
@@ -15,22 +18,30 @@ class InstallCommand extends Command
     {
         $this->comment('Publishing NativePHP Service Provider...');
         $this->callSilent('vendor:publish', ['--tag' => 'nativephp-provider']);
+        $this->callSilent('vendor:publish', ['--tag' => 'nativephp-config']);
+
 
         if ($this->option('force') || $this->confirm('Would you like to install the NativePHP NPM dependencies?', true)) {
-
             $this->comment('Installing NPM dependencies (This may take a while)...');
 
-            $this->option('installer') === 'yarn'
-                ? $this->installYarnDependencies()
-                : ($this->option('installer') === 'pnpm'
-                    ? $this->installPnpmDependencies()
-                    : $this->installNpmDependencies()
-            );
+            switch ($this->option('installer')) {
+                case 'yarn':
+                    $this->info('Installing NPM dependencies using the yarn package manager...');
+                    $this->installYarnDependencies();
+                    break;
+                case 'pnpm':
+                    $this->info('Installing NPM dependencies using the pnpm package manager...');
+                    $this->installPnpmDependencies();
+                    break;
+                default:
+                    $this->info('Installing NPM dependencies using the npm package manager...');
+                    $this->installNpmDependencies();
+            }
 
             $this->output->newLine();
         }
 
-        if (!$this->option('force') && $this->confirm('Would you like to start the NativePHP development server', false)) {
+        if (! $this->option('force') && $this->confirm('Would you like to start the NativePHP development server', false)) {
             $this->call('native:serve');
         }
 
@@ -39,22 +50,7 @@ class InstallCommand extends Command
 
     protected function nativePhpPath()
     {
-        return realpath(__DIR__ . '/../../resources/js');
-    }
-
-    protected function installNpmDependencies()
-    {
-        $this->executeCommand('npm set progress=false && npm install', $this->nativePhpPath());
-    }
-
-    protected function installYarnDependencies()
-    {
-        $this->executeCommand('yarn install', $this->nativePhpPath());
-    }
-
-    protected function installPnpmDependencies()
-    {
-        $this->executeCommand('pnpm install', $this->nativePhpPath());
+        return realpath(__DIR__.'/../../resources/js');
     }
 
     protected function executeCommand($command, $path)
